docker_fullImage/rest_services_docker/requirements.txt
face_recognition==1.3.0
Flask_Limiter==1.4
glob2==0.7
imutils==0.5.3
pandas==1.1.2
Flask==1.1.2
tensorflow_gpu==2.3.1
dlib==19.21.0
pytesseract==0.3.6
numpy==1.18.5
opencv_python==4.4.0.44
requests==2.24.0
Flask_Caching==1.9.0
memory_profiler==0.57.0
PassportEye==2.0.0
locustio==0.999
Pillow==7.2.0
tensorflow==2.3.1
